Gamma correction
The reason for this is that almost all computer programs assume that an
image is “gamma corrected” before being written into an image file. This
means that the 0 to 1 values have some transform applied before being
stored as a byte. Images with data that are written without being
transformed are said to be in linear space, whereas images that are
transformed are said to be in gamma space. It is likely that the image
viewer you are using is expecting an image in gamma space, but we are
giving it an image in linear space. This is the reason why our image
appears inaccurately dark.